requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.1.1
instagrapi==2.0.0
requests==2.31.0
